Header Message and Edit Profile Design Changes
Changed the header messages which display user information (now opaque background with purple colour).
Changed the input field of the biography for edit profile to be a textarea and restyled it to fit in the desired space.
Buttons for submitting changes or deleting profile on edit profile were styled and moved to be more in line with site themes